requirements.txt
PyQt5
pywin32
requests
pillow
numpy
opencv-python
pyinstaller
beautifulsoup4
pytesseract
psutil